Seperate ThingsBoard into Http and non Http.
