+++ examples/0005-arduino-sim900_send_telemetry_http/0005-arduino-sim900_send_telemetry_http.ino
@@ -18,7 +18,7 @@
 
 #include <TinyGsmClient.h>
 #include <SoftwareSerial.h>
-#include <ThingsBoard.h>
+#include <ThingsBoardHttp.h>
 
 
 // Your GPRS credentials
