+++ src/ThingsBoardDefaultLogger.cpp
@@ -1,8 +1,8 @@
 // Header include.
-#include "ThingsBoardDefaultLogger.h"
+#include <ThingsBoardDefaultLogger.h>
 
 // Library include.
-#include "HardwareSerial.h"
+#include <HardwareSerial.h>
 
 void ThingsBoardDefaultLogger::log(const char *msg) {
   Serial.print(F("[TB] "));
