+++ src/ThingsBoardDefaultLogger.h
@@ -1,5 +1,11 @@
-#ifndef THINGSBOARD_DEFAULT_LOGGER
-#define THINGSBOARD_DEFAULT_LOGGER
+/*
+  ThingsBoardDefaultLogger.h - Library API for sending data to the ThingsBoard
+  Based on PubSub MQTT library.
+  Created by Olender M. Oct 2018.
+  Released into the public domain.
+*/
+#ifndef Thingsboard_Default_Logger_h
+#define Thingsboard_Default_Logger_h
 
 class ThingsBoardDefaultLogger {
 public:
@@ -7,4 +13,4 @@ public:
     static void log(const char* msg);
 };
 
-#endif // THINGSBOARD_DEFAULT_LOGGER
+#endif // Thingsboard_Default_Logger_h
